Assistant checkpoint: Create basic landing page template
Assistant generated file changes:
- index.html: Create basic landing page template

---

User prompt:

I want to create a basic html site, that will be included as a part of a different site, therefore it comes with some limitations. The parent site is working on bootstrap, but all css must be added inline, and the site should use very limited jscript / jquery.  I have already a working project that meets this requirements, but now i need to add another page (it will work as a landing page). I can share with you the code of already working page which we could use as a referrence of some sort